backend/requirements.txt
fastapi
uvicorn[standard]
motor
redis[hiredis]
pyjwt
passlib[bcrypt]
python-dotenv
pydantic
email-validator
